seismic tx test utils (#156)
