+++ crates/seismic/primitives/Cargo.toml
@@ -28,7 +28,8 @@ alloy-eips = { workspace = true, features = ["k256"] }
 # op
 seismic-alloy-consensus.workspace = true
 alloy-rpc-types-eth = { workspace = true, optional = true }
-alloy-network = { workspace = true, optional = true }
+alloy-network = { workspace = true }
+seismic-alloy-network = { workspace = true }
 alloy-serde = { workspace = true, optional = true }
 seismic-revm.workspace = true
 
@@ -58,10 +59,8 @@ alloy-rpc-types.workspace = true
 k256.workspace = true
 enr = {workspace = true, features = ["rust-secp256k1"]}
 alloy-dyn-abi.workspace = true
-jsonrpsee.workspace = true
 reth-enclave.workspace = true
 seismic-alloy-rpc-types.workspace = true
-reth-chainspec.workspace = true
 
 [dev-dependencies]
 arbitrary.workspace = true
