+++ crates/seismic/txpool/src/transaction.rs
@@ -227,54 +227,54 @@ where
 pub trait SeismicPooledTx: PoolTransaction {}
 impl<T> SeismicPooledTx for T where T: PoolTransaction {}
 
-//#[cfg(test)]
-//mod tests {
-//    use crate::SeismicPooledTransaction;
-//    use alloy_consensus::transaction::Recovered;
-//    use alloy_eips::eip2718::Encodable2718;
-//    use alloy_primitives::{PrimitiveSignature as Signature, TxKind, U256};
-//    use reth_primitives_traits::transaction::error::InvalidTransactionError;
-//    use reth_provider::test_utils::MockEthProvider;
-//    use reth_seismic_chainspec::SEISMIC_MAINNET;
-//    use reth_seismic_evm::SeismicEvmConfig;
-//    use reth_seismic_primitives::test_utils::get_signed_seismic_tx;
-//    use reth_transaction_pool::{
-//        blobstore::InMemoryBlobStore, error::InvalidPoolTransactionError,
-//        validate::EthTransactionValidatorBuilder, TransactionOrigin, TransactionValidationOutcome,
-//    };
-//
-//    #[tokio::test]
-//    async fn validate_seismic_transaction() {
-//        // setup validator
-//        let client = MockEthProvider::default().with_chain_spec(SEISMIC_MAINNET.clone());
-//        let validator = EthTransactionValidatorBuilder::new(client)
-//            .no_shanghai()
-//            .no_cancun()
-//            .build(InMemoryBlobStore::default());
-//
-//        // check that a SeismicTypedTransaction::Seismic is valid
-//        let origin = TransactionOrigin::External;
-//        let signer = Default::default();
-//        let signed_seismic_tx = get_signed_seismic_tx();
-//        let signed_recovered = Recovered::new_unchecked(signed_seismic_tx, signer);
-//        let len = signed_recovered.encode_2718_len();
-//        let pooled_tx: SeismicPooledTransaction =
-//            SeismicPooledTransaction::new(signed_recovered, len);
-//
-//        let outcome = validator.validate_one(origin, pooled_tx);
-//
-//        match outcome {
-//            TransactionValidationOutcome::Invalid(
-//                _,
-//                InvalidPoolTransactionError::Consensus(InvalidTransactionError::InsufficientFunds(
-//                    _,
-//                )),
-//            ) => {
-//                // expected since the client (MockEthProvider) state does not have funds for any
-//                // accounts account balance is one of the last things checked in
-//                // validate_one, so getting that far good news
-//            }
-//            _ => panic!("Did not get expected outcome, got: {:?}", outcome),
-//        }
-//    }
-//}
+#[cfg(test)]
+mod tests {
+    use crate::SeismicPooledTransaction;
+    use alloy_consensus::transaction::Recovered;
+    use alloy_eips::eip2718::Encodable2718;
+    use alloy_primitives::{PrimitiveSignature as Signature, TxKind, U256};
+    use reth_primitives_traits::transaction::error::InvalidTransactionError;
+    use reth_provider::test_utils::MockEthProvider;
+    use reth_seismic_chainspec::SEISMIC_MAINNET;
+    use reth_seismic_evm::SeismicEvmConfig;
+    use reth_seismic_primitives::test_utils::get_signed_seismic_tx;
+    use reth_transaction_pool::{
+        blobstore::InMemoryBlobStore, error::InvalidPoolTransactionError,
+        validate::EthTransactionValidatorBuilder, TransactionOrigin, TransactionValidationOutcome,
+    };
+
+    #[tokio::test]
+    async fn validate_seismic_transaction() {
+        // setup validator
+        let client = MockEthProvider::default().with_chain_spec(SEISMIC_MAINNET.clone());
+        let validator = EthTransactionValidatorBuilder::new(client)
+            .no_shanghai()
+            .no_cancun()
+            .build(InMemoryBlobStore::default());
+
+        // check that a SeismicTypedTransaction::Seismic is valid
+        let origin = TransactionOrigin::External;
+        let signer = Default::default();
+        let signed_seismic_tx = get_signed_seismic_tx();
+        let signed_recovered = Recovered::new_unchecked(signed_seismic_tx, signer);
+        let len = signed_recovered.encode_2718_len();
+        let pooled_tx: SeismicPooledTransaction =
+            SeismicPooledTransaction::new(signed_recovered, len);
+
+        let outcome = validator.validate_one(origin, pooled_tx);
+
+        match outcome {
+            TransactionValidationOutcome::Invalid(
+                _,
+                InvalidPoolTransactionError::Consensus(InvalidTransactionError::InsufficientFunds(
+                    _,
+                )),
+            ) => {
+                // expected since the client (MockEthProvider) state does not have funds for any
+                // accounts account balance is one of the last things checked in
+                // validate_one, so getting that far good news
+            }
+            _ => panic!("Did not get expected outcome, got: {:?}", outcome),
+        }
+    }
+}
